Refactor communication/low_level

+++ src/communication/low_level/ConnectionSimulatorFlakey.cpp
@@ -8,7 +8,7 @@ ConnectionSimulatorFlakey::ConnectionSimulatorFlakey(double error_rate)
 
 uint8_t ConnectionSimulatorFlakey::Filter(uint8_t input) {
   if (rand() % 100000 < error_rate * 100000) {
-    return rand() % 256;
+    return static_cast<uint8_t>(rand() % 256);
   } else {
     return input;
   }
